requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
